tg/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-telegram-bot==20.7
pymysql==1.1.0
sqlalchemy==2.0.37
python-dotenv==0.21.1
typing_extensions==4.12.2
Werkzeug==3.1.3